A    https://scm.opendap.org/svn/branch/bes/dap4

+++ dispatch/BESScrub.cc
@@ -18,7 +18,7 @@
 // 
 // You should have received a copy of the GNU Lesser General Public
 // License along with this library; if not, write to the Free Software
-// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 //
 // You can contact University Corporation for Atmospheric Research at
 // 3080 Center Green Drive, Boulder, CO 80301
@@ -96,7 +96,7 @@ BESScrub::pathname_ok(const string &path, bool strict)
         
     string::size_type len = path.length() ;
     int ret = name.match( path.c_str(), len ) ;
-    if( ret != len )
+    if( ret != static_cast<int>(len) )
         return false ;
     return true ;
 }
